Refactor recording, visualization, and device capabilities structures
Streamlined and relocated interfaces for better modularity (`RecordingConfig` and `DeviceCapabilities`). Standardized service imports, enhanced documentation across services and components, and renamed `OptimizedCanvasRenderer` for clarity. Consolidated `VisualizationBar` and related interfaces into reusable entities. Removed unused methods, deprecated properties, and improved typing consistency for audio and recording workflows.
