+++ src/app/chat-ui/chat-ui-inputfield/audio-processing/audio-level-smoother.ts
@@ -1,5 +1,15 @@
 import { AUDIO_VISUALIZATION_CONFIG } from './audio-constants';
 
+
+/**
+ * The `AudioLevelSmoother` class provides functionality to smooth audio level data,
+ * allowing for gradual changes in the displayed levels to improve visualization
+ * stability. It uses an attack and release mechanism to manage quick rises and slower falls
+ * in audio level values.
+ *
+ * This class is primarily designed for use in audio visualization systems and ensures
+ * consistency and smooth transitions in audio level representation.
+ */
 export class AudioLevelSmoother {
   private currentLevel = 0;
   private readonly ATTACK = AUDIO_VISUALIZATION_CONFIG.ATTACK_RATE;
