--- src/app/chat-ui/chat-ui-inputfield/models/recording-config.interface.ts
@@ -1,12 +1,0 @@
-export interface RecordingConfig {
-  isHoldToRecord: boolean;
-  maxDuration: number; // in seconds
-  mimeType?: string;
-  audioConstraints?: MediaTrackConstraints;
-}
-
-export interface RecordingResult {
-  blob: Blob;
-  duration: number; // in seconds
-  mimeType: string;
-}
