--- src/app/chat-ui/chat-ui-inputfield/visualizers/audio-visualizer.interface.ts
@@ -1,11 +1,0 @@
-export interface AudioVisualizer {
-  start(): void;
-  stop(): void;
-  updateDimensions(width: number, height: number): void;
-}
-
-export interface VisualizationBar {
-  height: number;
-  timestamp: number;
-  x: number;
-}
