+++ src/app/data/objects/device-capabilities.ts
@@ -1,0 +1,7 @@
+export interface DeviceCapabilities {
+  hasCamera: boolean;
+  hasMultipleCameras: boolean;
+  hasGeolocation: boolean;
+  hasAudioInput: boolean;
+  isMobile: boolean;
+}
